+++ src/Utility/Exception/TransferException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Webthink\MonologSlack\Utility\Exception;
+
+/**
+ * This is the exception that will be throw in cases when a client could not communicate with slack.
+ */
+class TransferException extends \Exception
+{
+}