+++ tests/Unit/Handler/SlackWebhookHandlerTest.php
@@ -2,19 +2,19 @@
 
 namespace Webthink\MonologSlack\Test\Unit\Handler;
 
-use GuzzleHttp\Client;
-use GuzzleHttp\Exception\RequestException;
-use GuzzleHttp\Psr7\Request;
 use Monolog\Formatter\LineFormatter;
 use Monolog\Logger;
+use PHPUnit\Framework\MockObject\MockObject;
 use Webthink\MonologSlack\Formatter\SlackShortAttachmentFormatter;
 use Webthink\MonologSlack\Handler\SlackWebhookHandler;
 use Webthink\MonologSlack\Test\Unit\TestCase;
+use Webthink\MonologSlack\Utility\ClientInterface;
+use Webthink\MonologSlack\Utility\Exception\TransferException;
 
 class SlackWebhookHandlerTest extends TestCase
 {
     /**
-     * @var Client|\PHPUnit_Framework_MockObject_MockObject
+     * @var ClientInterface|MockObject
      */
     private $client;
 
@@ -29,7 +29,7 @@ class SlackWebhookHandlerTest extends TestCase
     public function setUp()
     {
         parent::setUp();
-        $this->client = $this->getMockBuilder(Client::class)->disableOriginalConstructor()->getMock();
+        $this->client = $this->getMockBuilder(ClientInterface::class)->getMock();
         $this->handler = new SlackWebhookHandler('www.dummy.com', null, 'rotating_light', Logger::ERROR, true, $this->client);
     }
 
@@ -48,40 +48,33 @@ class SlackWebhookHandlerTest extends TestCase
      */
     public function handlerWillHandleTheRecord()
     {
-        $this->client->expects($this->once())->method('request')
-            ->with(
-                'post',
-                'www.dummy.com',
-                $this->callback(function ($value) {
-                    if (!is_array($value)) {
-                        return false;
-                    }
+        $this->client->expects($this->once())->method('send')
+            ->with('www.dummy.com', $this->callback(function ($value) {
+                if (!is_array($value)) {
+                    return false;
+                }
 
-                    $this->assertSame(':rotating_light:', $value['json']['icon_emoji']);
-                    $this->assertStringEndsWith("test.CRITICAL: test [] []\n", $value['json']['text']);
-                    return true;
-                })
-            )
+                $this->assertSame(':rotating_light:', $value['icon_emoji']);
+                $this->assertStringEndsWith("test.CRITICAL: test [] []\n", $value['text']);
+                return true;
+            }))
             ->willReturn(null);
         $this->handler->handle($this->getRecord(Logger::CRITICAL));
     }
 
     /**
      * @test
+     * @expectedException \Webthink\MonologSlack\Utility\Exception\TransferException
      */
     public function clientWillThrowExceptionButHandlerWillFailSilently()
     {
-        $this->client->expects($this->once())->method('request')
-            ->with(
-                'post',
-                'www.dummy.com',
-                $this->callback(function (array $value) {
-                    $this->assertSame(':rotating_light:', $value['json']['icon_emoji']);
-                    $this->assertStringEndsWith("test.CRITICAL: test [] []\n", $value['json']['text']);
-                    return true;
-                })
-            )
-            ->willThrowException(new RequestException('Bad Request', new Request('get', 'www.dummy.com')));
+        $this->client->expects($this->once())->method('send')
+            ->with('www.dummy.com', $this->callback(function (array $value) {
+                $this->assertSame(':rotating_light:', $value['icon_emoji']);
+                $this->assertStringEndsWith("test.CRITICAL: test [] []\n", $value['text']);
+                return true;
+            }))
+            ->willThrowException(new TransferException('Bad Request', 400));
         $this->handler->handle($this->getRecord(Logger::CRITICAL));
     }
 
@@ -90,7 +83,7 @@ class SlackWebhookHandlerTest extends TestCase
      */
     public function handlerDoesNotHandleTheRecord()
     {
-        $this->client->expects($this->never())->method('request');
+        $this->client->expects($this->never())->method('send');
         $this->handler->handle($this->getRecord());
     }